refactor: 에러 핸들링 로직 기능성 컴포넌트로 분리 (QueryClientDefaultOptionProvider) #358
